requirements.txt
hikari[server]
hikari-crescent
tortoise-orm[asyncpg]